refactor: layout page with dynamic router with Optional Catch-all Segments [[...slug]]
