+++ src/layouts/body/header.type.ts
@@ -1,0 +1,5 @@
+export interface DynamicBodyProps {
+  item: Record<string, any>;
+  systemConfig: Record<string, any>;
+  slug?: string;
+}
