+++ src/layouts/footer/footer.layout.tsx
@@ -1,18 +1,13 @@
 import React from 'react';
-import { FooterTheme02 } from '../../next-core-ui/share-account/src/layouts/footer/footer.layout';
-
-export interface DynamicFooterProps {
-  item: Record<string, any>;
-  systemConfig: Record<string, any>;
-  slug?: string;
-}
+import { DynamicFooterProps } from './footer.type';
+import { FooterTheme01 } from '../../next-core-ui';
 
 export const DynamicFooter: React.FC<DynamicFooterProps> = (props) => {
   const { item, systemConfig, slug } = props;
 
   switch (item?.type) {
     case 'FOOTER_THEME_01':
-      return <FooterTheme02 />;
+      return <FooterTheme01 />;
     case 'FOOTER_NONE':
       return <div className="hidden" />;
     default:
