+++ src/layouts/footer/footer.type.ts
@@ -1,0 +1,5 @@
+export interface DynamicFooterProps {
+  item: Record<string, any>;
+  systemConfig: Record<string, any>;
+  slug?: string;
+}
