+++ src/layouts/header/header.type.ts
@@ -1,0 +1,6 @@
+export interface DynamicHeaderProps {
+  item: Record<string, any>;
+  systemConfig: Record<string, any>;
+  wuiHeaderContent: Record<string, any>;
+  slug?: string;
+}
