+++ src/layouts/header/index.ts
@@ -1,1 +1,2 @@
-export * from './dynamic-header.component';
+export * from './header.layout';
+export * from './header.type';
