+++ src/next-core-ui/share-account/src/layouts/footer/footer.type.ts
@@ -1,1 +1,1 @@
-export interface IFooterTheme02Props {}
+export interface IFooterTheme01Props {}
