+++ src/next-core-ui/share-account/src/layouts/header/header.layout.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import { IHeaderTheme02Props } from './header.type';
 import { MenuDesktop, MenuMobile } from '../../views';
 import { useDeviceSizes } from '../../../../../hooks';
+import { IHeaderTheme01Props } from './header.type';
 
-export const HeaderTheme02: React.FC<IHeaderTheme02Props> = (props) => {
+export const HeaderTheme01: React.FC<IHeaderTheme01Props> = (props) => {
   const { wuiHeaderContent } = props;
 
   const isDevice = useDeviceSizes();
