+++ src/next-core-ui/share-account/src/layouts/header/header.type.ts
@@ -1,3 +1,3 @@
-export interface IHeaderTheme02Props {
+export interface IHeaderTheme01Props {
   wuiHeaderContent?: Record<string, any>;
 }
