+++ src/pages/[[...slug]]/index.tsx
@@ -1,20 +1,21 @@
 import React, { useMemo } from 'react';
 import Head from 'next/head';
 import { observer } from 'mobx-react';
+import { DynamicHeader } from '../../layouts/header/header.layout';
+import { DynamicFooter } from '../../layouts/footer/footer.layout';
 import { useRouter } from 'next/router';
 import { GetServerSideProps } from 'next';
-import { IDynamicPageProps } from '..';
-import { DynamicHeader, DynamicBody, DynamicFooter } from '../../../layouts';
-import { rootLayoutConfig } from '../../../mocks';
-import { fetchUiContentApi } from '../../../store/store-ui-content/api';
+import { rootLayoutConfig } from '../../mocks';
+import { fetchUiContentApi } from '../../store/store-ui-content/api';
+import { DynamicBody } from '../../layouts/body/body.layout';
 
-export interface IDynamicPageChildProps {
-  systemConfig: Record<string, any>;
-  appShell: Record<string, any>;
-  wuiHeaderContent: Record<string, any>;
+export interface IDynamicPageProps {
+  systemConfig?: Record<string, any>;
+  appShell?: Record<string, any>;
+  wuiHeaderContent?: Record<string, any>;
 }
 
-const DynamicPageChild: React.FC<IDynamicPageChildProps> = observer((props) => {
+const DynamicPage: React.FC<IDynamicPageProps> = observer((props) => {
   const { systemConfig, appShell, wuiHeaderContent } = props;
 
   const router = useRouter();
@@ -22,11 +23,8 @@ const DynamicPageChild: React.FC<IDynamicPageChildProps> = observer((props) => {
   const slug = useMemo(() => {
     const str = router?.query?.slug;
 
-    if (!str) return '';
+    if (!str) return 'HOME';
 
-    if (typeof str === 'string') {
-      return str.toLocaleUpperCase();
-    }
     return str[0].toLocaleUpperCase();
   }, [router?.query?.slug]);
 
@@ -40,12 +38,11 @@ const DynamicPageChild: React.FC<IDynamicPageChildProps> = observer((props) => {
   return (
     <>
       <Head>
-        <meta property="og:title" content={seoHomePage?.metaSeo || 'DQH Architectures - Trang Liên hệ'} />
-        <meta property="og:url" content={seoHomePage?.metaUrl || 'https://dqharchitects.vn/contact'} />
+        <meta property="og:title" content={seoHomePage?.metaSeo} />
+        <meta property="og:url" content={seoHomePage?.metaUrl} />
         <meta property="og:image" />
-        <title>{seoHomePage?.title || 'DQH Architectures - Liên hệ'}</title>
+        <title>{seoHomePage?.title}</title>
       </Head>
-
       <DynamicHeader
         item={contentPage?.header}
         systemConfig={systemConfig}
@@ -58,7 +55,7 @@ const DynamicPageChild: React.FC<IDynamicPageChildProps> = observer((props) => {
   );
 });
 
-export default DynamicPageChild;
+export default DynamicPage;
 
 export const getServerSideProps = (async () => {
   const systemConfig: Record<string, any> = (await fetchUiContentApi({ contentId: 'ui-config' })) || {};
